update, remove wating time, add more signal

+++ main.c
@@ -1,27 +1,37 @@
 #include <unistd.h>
 #include <sys/wait.h>
 #include <stdio.h>
-#include <stdlib.h>
 #include <errno.h>
 #include <string.h>
-#include <signal.h>
-#include <time.h>
 
 static char *prog_name;
 
 static int cpid = 0;
 
-static void sig_forwarder(int signo) {
-	kill(cpid, signo);
-}
-
 static void exit_error(char *msg) {
 	fprintf(stderr, "[ERROR %s] %s\n", prog_name, msg);
-	exit(1);
+	_exit(1);
 }
 
 static void exit_errno() {
 	exit_error(strerror(errno));
+}
+
+static int siglist[] = {SIGHUP, SIGINT, SIGQUIT, SIGTERM, SIGUSR1, SIGUSR2};
+
+static void forward_signal(int signo) {
+	if (cpid != 0) kill(cpid, signo);
+}
+
+static void register_handler() {
+	struct sigaction act;
+	memset(&act, 0, sizeof(struct sigaction));
+	act.sa_handler = forward_signal;
+	act.sa_flags = SA_RESTART;
+
+	for (int i = 0; i < sizeof(siglist) / sizeof(siglist[0]); ++i) {
+		if (sigaction(siglist[i], &act, NULL) == -1) exit_errno();
+	}
 }
 
 int main(int argc, char *argv[]) {
@@ -29,33 +39,25 @@ int main(int argc, char *argv[]) {
 
 	if (getpid() != 1) exit_error("PID is not 1");
 
-	if (argc == 1) exit_error("require at least 1 argument");
+	if (argc == 1) exit_error("No command to be executed, require at least 1 argument");
 
 	cpid = fork();
 	if (cpid == -1) exit_errno();
 
 	if (cpid) {
-
-		if (signal(SIGINT, sig_forwarder) == SIG_ERR) exit_errno();
-		if (signal(SIGTERM, sig_forwarder) == SIG_ERR) exit_errno();
+		register_handler();
 
 		int cstatus = 0;
-		int count = 5;
-		const struct timespec tenmili = { .tv_sec = 0, .tv_nsec = 10000000 };
 
 		for (;;) {
 			int status;
 			int pid = wait(&status);
 
 			if (pid == -1) {
-				if (errno == ECHILD) {
-					nanosleep(&tenmili, NULL);
-					count--;
-					if (count <= 0) return WEXITSTATUS(cstatus);
-				} else {
-					exit_errno();
-				}
+				if (errno == ECHILD) return WEXITSTATUS(cstatus);
+				else exit_errno();
 			} else if (pid == cpid) {
+				cpid = 0;
 				cstatus = status;
 			}
 		}
